raytracer::objects

Module circle

Circle object initialization and behaviors

This module defines the circle object type used throughout the raytracer. Circle objects serve as both standalone entities and as the visual base for other object types like emitters. The module provides implementation of the core Drawable and Movable traits.

author: Zhean Ganituen (zrygan) last updated: April 16, 2025

Structs

ObjectCircle

Represents a basic circle object in the raytracer.